backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
cassandra-driver==3.29.2